Move Gerrit specific classes to package with Gerrit provider
Gerrit specific classes are easier to find when they are next to
other Gerrit classes.
