+++ src/test/java/com/tw/go/plugin/provider/gerrit/GerritProviderTest.java
@@ -1,6 +1,6 @@
 package com.tw.go.plugin.provider.gerrit;
 
-import com.tw.go.plugin.setting.GerritPluginSettings;
+import com.tw.go.plugin.provider.gerrit.settings.GerritPluginSettings;
 import com.tw.go.plugin.util.AuthenticationType;
 import com.tw.go.plugin.util.HTTPClient;
 import org.junit.Before;
